# shareplate-backend/swagger.yaml
...
openapi: 3.0.3
info:
    title: SharePlate API Doc
    description: This is the RESTful API Docs for SDA for team 2
    version: 0.0.2
servers:
    - url: http://localhost:3000/api/
...
paths:
...
    "/api/shops/me":
        get:
            tags:
                - Shop
            summary: Get shop by token
            description: Obtain shop information by token
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Shop"
                "400":
                    description: Bad Request
                "401":
                    description: Unauthorized
                "500":
                    description: Internal Server Error
    "/api/shops":
        post:
            tags:
                - Shop
            summary: Create shop
            description: Create a new shop.
            requestBody:
                description: username and password
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                username:
                                    type: string
                                password:
                                    type: string
                            example:
                                username: Starbucks
                                password: starbucksPassword
            responses:
                "200":
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: string
                                    token:
                                        type: string
                                example:
                                    id: 632b58c9878188af6c4d8075
                                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0MWE4ZTRjNDhkNmQ1ZGYzMzM3MWJmZSIsInR5cGUiOjEsImlhdCI6MTY3OTQ2MTk2NCwiZXhwIjoxNzEwOTk3OTY0fQ.G-TpWhojU-9zm97XXS35Mgp2ed4mWoYWBPZQZRaVeto
                "400":
                    description: Bad Request
                "401":
                    description: Unauthorized
                "500":
                    description: Internal Server Error
    "/api/shops/sessions":
        post:
            tags:
                - Shop
            summary: Login
            description: Verify username and password, then return jwt token
            requestBody:
                description: username and password
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                username:
                                    type: string
                                password:
                                    type: string
                            example:
                                username: Starbucks
                                password: starbucksPassword
            responses:
                "200":
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: string
                                    token:
                                        type: string
                                example:
                                    id: 632b58c9878188af6c4d8075
                                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjMyYjU4Yzk4NzgxODhhZjZjNGQ4MDc1IiwiaWF0IjoxNTE2MjM5MDIyfQ.PTgD-8uBUOwxnkBwTn_A76LXfAF6xcR7JmsleHnSlDs
                "400":
                    description: Bad Request
                "401":
                    description: Unauthorized
                "500":
                    description: Internal Server Error
    "/api/shops/{shop_id}":
        get:
            tags:
                - Shop
            description: get shop by id
            parameters:
                - name: shop_id
                  in: path
                  required: true
                  schema:
                      type: string
                      example: 641ce90385ccdb788fee248a
            responses:
                "200":
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Shop"
                "500":
                    description: Internal Server Error
        patch:
            tags:
                - Shop
            description: modify shop information
            parameters:
                - name: shop_id
                  in: path
                  required: true
                  schema:
                      type: string
                      example: 641ce90385ccdb788fee248a
            requestBody:
                description: updated shop information
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Shop"
            responses:
                "200":
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Shop"
                "500":
                    description: Internal Server Error
...
components:
    schemas:
...
        Shop:
            type: object
            required:
                - _id
                - shopName
                - location
            properties:
                _id:
                    type: string
                    description: Shop ID
                    example: 641a8e4c48d6d5df33371bfe
                shopName:
                    type: string
                    description: Shop name
                    example: Shop A
                location:
                    type: object
                    properties:
                        coordinates:
                            type: array
                            items:
                                type: number
                            description: The coordinates of the shop location (longitude, latitude)
                            example: [-122.421778, 37.783943]
                    description: The location of the shop
                description:
                    type: string
                    description: Shop description
                    example: Best coffee in town!
                shopType:
                    type: array
                    items:
                        type: string
                    description: The types of the shop
                    example: [coffee, bakery]
                followers:
                    type: array
                    items:
                        type: string
                        format: objectId
                    description: The IDs of the users who follow this shop
                menu:
                    type: array
                    items:
                        type: object
                        properties:
                            name:
                                type: string
                                description: Name of the menu item
                                example: Cappuccino
                            price:
                                type: number
                                format: double
                                description: Price of the menu item
                                example: 3.99
                    description: The menu of the shop
                available:
                    type: boolean
                    description: The availability status of the shop
                    example: false
                reviews:
                    type: array
                    items:
                        type: object
                        properties:
                            _id:
                                type: string
                                description: Review ID
                                example: 610dc5ba987e7f3b18d8d994
                            reviewer:
                                type: string
                                description: Reviewer's user ID
                                example: 610dc52e987e7f3b18d8d990
                                format: objectId
                            createdTime:
                                type: string
                                description: Timestamp of the review creation
                                example: "2022-03-23T08:18:19.961Z"
                                format: date-time
                            content:
                                type: string
                                description: Content of the review
                                example: This is a great shop!
                            rating:
                                type: integer
                                description: Rating of the review
                                example: 5
                                minimum: 0
                                maximum: 5
                    description: The reviews of the shop
                __v:
                    type: integer
                    description: The version of the shop document
                    example: 0
            example:
                _id: 641a8e4c48d6d5df33371bfe
                shopName: Shop A
                location:
                    coordinates: [-122.421778, 37.783943]
                description: Best coffee in town!
                shopType: [coffee, bakery]
                followers: []
                menu:
                    - name: Cappuccino
                      price: 3.99
                available: false
                reviews: []
                __v: 0
...